:sparkles: feat: GamePulse DTO를 Demo.Web으로 마이그레이션 및 FluentValidation 추가
