+++ NET10/Demo.Web/Demo.Web.csproj
@@ -8,6 +8,7 @@
 
     <ItemGroup>
         <PackageReference Include="FastEndpoints" Version="7.0.1" />
+        <PackageReference Include="FluentValidation" Version="12.0.0" />
         <PackageReference Include="LiteBus" Version="3.0.0" />
         <PackageReference Include="LiteBus.Commands" Version="3.0.0" />
         <PackageReference Include="LiteBus.Commands.Abstractions" Version="3.0.0" />
