+++ NET10/GamePulse.Test/GamePulse.Test.csproj
@@ -23,6 +23,7 @@
     <ProjectReference Include="..\GamePulse\GamePulse.csproj" />
     <ProjectReference Include="..\Demo.Application\Demo.Application.csproj" />
     <ProjectReference Include="..\Demo.Infra\Demo.Infra.csproj" />
+    <ProjectReference Include="..\Demo.Web\Demo.Web.csproj" />
   </ItemGroup>
 
   <ItemGroup>
