+++ NET10/GamePulse/EndPoints/User/Create/CreateEndpointV1.cs
@@ -1,5 +1,5 @@
 using FastEndpoints;
-using GamePulse.DTO;
+using Demo.Web.DTO;
 using GamePulse.Processors;
 using OpenTelemetry.Trace;
