+++ NET10/docs/clean-architecture-migration/task-003-dto-migration.md
@@ -1,0 +1,243 @@
+# Task 003: GamePulse DTO를 Demo.Web으로 마이그레이션
+
+## 개요
+CleanArchitecture 적용을 위해 GamePulse/DTO 폴더의 모든 파일들을 Demo.Web/DTO로 이동시키는 작업을 수행했습니다. DTO(Data Transfer Object)는 Web/Presentation 레이어에 속하므로 Demo.Web 프로젝트가 적절한 위치입니다.
+
+## 작업 내용
+
+### 1. 파일 이동 및 네임스페이스 변경
+
+#### 이동된 파일들
+- `GamePulse/DTO/LoginRequest.cs` → `Demo.Web/DTO/LoginRequest.cs`
+- `GamePulse/DTO/LoginRequestValidator.cs` → `Demo.Web/DTO/LoginRequestValidator.cs`
+- `GamePulse/DTO/MyRequest.cs` → `Demo.Web/DTO/MyRequest.cs`
+- `GamePulse/DTO/MyRequestValidator.cs` → `Demo.Web/DTO/MyRequestValidator.cs`
+- `GamePulse/DTO/MyResponse.cs` → `Demo.Web/DTO/MyResponse.cs`
+
+#### 네임스페이스 변경
+- `GamePulse.DTO` → `Demo.Web.DTO`
+
+### 2. 코드 개선사항
+
+#### 한국어 주석 및 문서화
+- 모든 XML 문서 주석을 한국어로 번역
+- 클래스와 속성에 대한 명확한 설명 추가
+- 유효성 검사 메시지를 한국어로 현지화
+
+#### 코드 품질 향상
+- 불필요한 using 문 제거
+- 일관된 네이밍 규칙 적용
+- 4개 공백 들여쓰기 유지
+
+### 3. 의존성 업데이트
+
+#### Demo.Web 프로젝트
+- FluentValidation 패키지 추가: `FluentValidation` Version="12.0.0"
+
+#### GamePulse.Test 프로젝트
+- Demo.Web 프로젝트 참조 추가
+
+#### 참조 업데이트된 파일들
+- `GamePulse/EndPoints/User/Create/CreateEndpointV1.cs`
+- `GamePulse/EndPoints/User/Create/CreateEndpointV2.cs`
+- `GamePulse/EndPoints/Login/LoginEndpointV1.cs`
+- `GamePulse.Test/DTO/MyRequestValidatorTests.cs`
+- `GamePulse.Test/DTO/LoginRequestValidatorTests.cs`
+
+## CleanArchitecture 원칙 적용
+
+### Web/Presentation Layer (Demo.Web)
+- **DTO 클래스들**: 외부와의 데이터 교환을 위한 객체들
+- **Validator 클래스들**: 입력 데이터 유효성 검사 로직
+
+### 레이어 분리의 이점
+1. **관심사 분리**: 데이터 전송 객체가 적절한 레이어에 위치
+2. **의존성 관리**: Web 레이어의 DTO가 다른 레이어에 영향을 주지 않음
+3. **재사용성**: DTO와 Validator가 Web 컨텍스트에서 재사용 가능
+
+## 파일별 상세 변경사항
+
+### LoginRequest.cs
+```csharp
+// 변경 전
+namespace GamePulse.DTO;
+
+/// <summary>
+/// Login 요청
+/// </summary>
+public class LoginRequest
+{
+    /// <summary>
+    /// 로그인 사용자명
+    /// </summary>
+    public string Username { get; set; } = string.Empty;
+    /// <summary>
+    /// 로그인 비밀번호
+    /// </summary>
+    public string Password { get; set; } = string.Empty;
+}
+
+// 변경 후
+namespace Demo.Web.DTO;
+
+/// <summary>
+/// 로그인 요청 DTO
+/// </summary>
+public class LoginRequest
+{
+    /// <summary>
+    /// 로그인 사용자명
+    /// </summary>
+    public string Username { get; set; } = string.Empty;
+
+    /// <summary>
+    /// 로그인 비밀번호
+    /// </summary>
+    public string Password { get; set; } = string.Empty;
+}
+```
+
+### LoginRequestValidator.cs
+```csharp
+// 주요 개선사항
+- 한국어 유효성 검사 메시지
+- 명확한 XML 문서 주석
+- 네임스페이스 변경: GamePulse.DTO → Demo.Web.DTO
+```
+
+### MyRequest.cs
+```csharp
+// 변경 전
+namespace GamePulse.DTO;
+
+/// <summary>
+/// 요청 예제
+/// </summary>
+
+// 변경 후
+namespace Demo.Web.DTO;
+
+/// <summary>
+/// 요청 예제 DTO
+/// </summary>
+```
+
+### MyRequestValidator.cs
+```csharp
+// 주요 개선사항
+- 한국어 유효성 검사 메시지 유지
+- XML 문서 주석 개선
+- 네임스페이스 변경
+```
+
+### MyResponse.cs
+```csharp
+// 변경 전
+namespace GamePulse.DTO;
+
+/// <summary>
+/// Response Exam
+/// </summary>
+
+// 변경 후
+namespace Demo.Web.DTO;
+
+/// <summary>
+/// 응답 예제 DTO
+/// </summary>
+```
+
+## 프로젝트 의존성 구조
+
+### 업데이트된 의존성
+```
+GamePulse → Demo.Web (기존 참조 유지)
+GamePulse.Test → Demo.Web (새로 추가)
+Demo.Web → FluentValidation (새로 추가)
+```
+
+### 의존성 방향 (CleanArchitecture 준수)
+```
+GamePulse (Presentation) → Demo.Web (Presentation/Shared)
+GamePulse (Presentation) → Demo.Application (Business Logic)
+GamePulse (Presentation) → Demo.Infra (Infrastructure)
+Demo.Web (Presentation) → Demo.Application (Business Logic)
+Demo.Web (Presentation) → Demo.Infra (Infrastructure)
+Demo.Infra (Infrastructure) → Demo.Application (Business Logic)
+```
+
+## 검증 사항
+
+### 빌드 검증
+- [x] Demo.Web 프로젝트 빌드 성공
+- [x] GamePulse 프로젝트 빌드 성공
+- [x] GamePulse.Test 프로젝트 빌드 성공
+- [x] 전체 솔루션 빌드 성공
+
+### 기능 검증
+- [x] 네임스페이스 참조 업데이트 완료
+- [x] 유효성 검사기 정상 작동
+- [x] 테스트 코드 정상 실행
+
+### 코드 품질 검증
+- [x] 한국어 주석 및 문서화 완료
+- [x] 불필요한 using 문 제거
+- [x] 일관된 코딩 스타일 적용
+
+## 빌드 결과
+
+### 최종 빌드 상태
+```
+성공 빌드(14.8초) - 1개 경고
+✅ Demo.Application: 성공
+✅ Demo.Infra: 성공
+✅ Demo.Application.Tests: 성공
+✅ Demo.Web: 성공 (1개 경고 - JsonSerializableAttribute 관련, 기능에 영향 없음)
+✅ GamePulse: 성공
+✅ Demo.Web.IntegrationTests: 성공
+✅ Demo.Web.PerformanceTests: 성공
+✅ GamePulse.Test: 성공
+```
+
+### 경고 분석
+- **SYSLIB1224**: JsonSerializableAttribute 관련 경고
+- 기능에 영향을 주지 않는 컴파일러 최적화 관련 경고
+- 필요시 JsonSerializerContext 구현으로 해결 가능
+
+## 다음 단계
+
+1. **기능 테스트**: 로그인 및 사용자 생성 엔드포인트 테스트
+2. **통합 테스트**: DTO 유효성 검사 및 직렬화 테스트
+3. **추가 마이그레이션**: 다른 GamePulse 컴포넌트들의 CleanArchitecture 적용 검토
+
+## 파일 변경 요약
+
+### 생성된 파일
+- `Demo.Web/DTO/LoginRequest.cs`
+- `Demo.Web/DTO/LoginRequestValidator.cs`
+- `Demo.Web/DTO/MyRequest.cs`
+- `Demo.Web/DTO/MyRequestValidator.cs`
+- `Demo.Web/DTO/MyResponse.cs`
+
+### 수정된 파일
+- `Demo.Web/Demo.Web.csproj` (FluentValidation 패키지 추가)
+- `GamePulse.Test/GamePulse.Test.csproj` (Demo.Web 프로젝트 참조 추가)
+- `GamePulse/EndPoints/User/Create/CreateEndpointV1.cs` (네임스페이스 참조 업데이트)
+- `GamePulse/EndPoints/User/Create/CreateEndpointV2.cs` (네임스페이스 참조 업데이트)
+- `GamePulse/EndPoints/Login/LoginEndpointV1.cs` (네임스페이스 참조 업데이트)
+- `GamePulse.Test/DTO/MyRequestValidatorTests.cs` (네임스페이스 참조 업데이트)
+- `GamePulse.Test/DTO/LoginRequestValidatorTests.cs` (네임스페이스 참조 업데이트)
+
+### 삭제된 파일
+- `GamePulse/DTO/` 디렉토리 전체 삭제
+
+## 결론
+
+GamePulse/DTO의 모든 파일들이 CleanArchitecture 원칙에 따라 Demo.Web 프로젝트로 성공적으로 마이그레이션되었습니다. 이를 통해:
+
+1. **적절한 레이어 분리**: DTO가 Web/Presentation 레이어에 위치
+2. **코드 품질 향상**: 한국어 문서화 및 일관된 스타일 적용
+3. **의존성 관리 개선**: CleanArchitecture 원칙에 따른 의존성 구조
+4. **유지보수성 향상**: 명확한 책임 분리와 재사용성 증대
+
+모든 프로젝트가 성공적으로 빌드되어 DTO 마이그레이션 작업이 안정적으로 완료되었습니다.
